code refactoring controllers

+++ src/main/java/com/runapp/achievementservice/dto/request/AchievementRequest.java
@@ -17,7 +17,7 @@ public class AchievementRequest {
 
     private String description;
 
-    private int rarity_id;
+    private Long rarity_id;
 
     public AchievementModel ToAchievementModel(RarityModel rarityModel) {
         AchievementModel achievementModel = new AchievementModel();
