+++ src/main/java/com/runapp/achievementservice/repository/RarityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RarityRepository extends JpaRepository<RarityModel, Integer> {
+public interface RarityRepository extends JpaRepository<RarityModel, Long> {
 }
