+++ src/main/java/com/runapp/achievementservice/repository/TrainingRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface TrainingRepository extends JpaRepository<TrainingModel, Integer> {
+public interface TrainingRepository extends JpaRepository<TrainingModel, Long> {
     @Query("SELECT t FROM TrainingModel t ORDER BY t.averagePace ASC NULLS LAST")
     TrainingModel findTrainingWithMinAveragePace();
